broker/requirements.txt
paho-mqtt>=1.4.0
psycopg2-binary>=2.7.4
requests>=2.19.1
SQLAlchemy>=1.3.3
virtualenvwrapper>=4.8.4
pandas>=0.23.1
